x/mining/types/message_set_max_reward_pool_number_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stafihub/stafihub/testutil/sample"
	"github.com/stretchr/testify/require"
)

func TestMsgSetMaxRewardPoolNumber_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSetMaxRewardPoolNumber
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSetMaxRewardPoolNumber{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSetMaxRewardPoolNumber{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
